tests: move storage_service_for_tests impl to test_services.cc
Let's make it easier to find.
